# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: 微服务
  version: ""
host: localhost:8000
basePath: /api
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /user/get_captcha_image:
    post:
      tags:
      - User
      summary: 获取图形验证码
      description: 获取图形验证码
      operationId: User#GetCaptchaImage
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserGetCaptchaImageResponseBody'
            required:
            - errcode
            - errmsg
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UserGetCaptchaImageBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UserGetCaptchaImageInternalServerErrorResponseBody'
      schemes:
      - http
  /user/login_by_mobile:
    post:
      tags:
      - User
      summary: 使用短信验证码登录
      description: 使用短信验证码登录
      operationId: User#LoginBySmsCode
      parameters:
      - name: LoginBySmsCodeRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserLoginBySmsCodeRequestBody'
          required:
          - mobile
          - verify_code
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserLoginBySmsCodeResponseBody'
            required:
            - errcode
            - errmsg
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UserLoginBySmsCodeBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UserLoginBySmsCodeInternalServerErrorResponseBody'
      schemes:
      - http
  /user/login_by_username:
    post:
      tags:
      - User
      summary: 使用账号密码登录
      description: 使用账号密码登录
      operationId: User#LoginByUsername
      parameters:
      - name: LoginByUsernameRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserLoginByUsernameRequestBody'
          required:
          - username
          - password
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserLoginByUsernameResponseBody'
            required:
            - errcode
            - errmsg
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UserLoginByUsernameBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UserLoginByUsernameInternalServerErrorResponseBody'
      schemes:
      - http
  /user/send_sms_code:
    post:
      tags:
      - User
      summary: 发送短信验证码
      description: 发送短信验证码
      operationId: User#SendSmsCode
      parameters:
      - name: SendSmsCodeRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserSendSmsCodeRequestBody'
          required:
          - mobile
          - humanCode
          - captchaId
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserSendSmsCodeResponseBody'
            required:
            - errcode
            - errmsg
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UserSendSmsCodeBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UserSendSmsCodeInternalServerErrorResponseBody'
      schemes:
      - http
  /user/update_password:
    post:
      tags:
      - User
      summary: 修改登录密码
      description: |-
        修改登录密码

        **Required security scopes for jwt**:
          * `role:user`
      operationId: User#UpdatePassword
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: true
        type: string
      - name: UpdatePasswordRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserUpdatePasswordRequestBody'
          required:
          - old_password
          - new_password
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserUpdatePasswordResponseBody'
            required:
            - errcode
            - errmsg
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UserUpdatePasswordBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UserUpdatePasswordInternalServerErrorResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
definitions:
...
  SuccessResultResponseBody:
    title: 'Mediatype identifier: successresult; view=default'
    type: object
    properties:
      ok:
        type: boolean
        description: success
        example: true
    description: 成功信息 (default view)
    example:
      ok: true
    required:
    - ok
...
  UserGetCaptchaImageBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GetCaptchaImage_bad_request_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserGetCaptchaImageInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GetCaptchaImage_internal_server_error_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserGetCaptchaImageResponseBody:
    title: UserGetCaptchaImageResponseBody
    type: object
    properties:
      data:
        $ref: '#/definitions/CaptchaResponseBody'
      errcode:
        type: integer
        description: 错误码
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
    example:
      data:
        captchaId: Debitis velit inventore.
        image: Facilis architecto voluptatem magni magni odit sit.
      errcode: 0
      errmsg: ""
    required:
    - errcode
    - errmsg
  UserLoginBySmsCodeBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: LoginBySmsCode_bad_request_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserLoginBySmsCodeInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: LoginBySmsCode_internal_server_error_Response_Body result type (default
      view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserLoginBySmsCodeRequestBody:
    title: UserLoginBySmsCodeRequestBody
    type: object
    properties:
      mobile:
        type: string
        description: 手机号
        example: guest
        minLength: 11
        maxLength: 11
      verify_code:
        type: string
        description: 短信验证码
        example: "123456"
        minLength: 4
        maxLength: 6
    example:
      mobile: guest
      verify_code: "123456"
    required:
    - mobile
    - verify_code
  UserLoginBySmsCodeResponseBody:
    title: UserLoginBySmsCodeResponseBody
    type: object
    properties:
      data:
        $ref: '#/definitions/SessionResponseBody'
      errcode:
        type: integer
        description: 错误码
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
    example:
      data:
        cookie: Eos optio ea voluptas voluptatem magnam nostrum.
        credentials:
          expires_in: 25200
          token: eyJhbGciOiJIUz...
        user:
          id: Ipsam magni quia minus officiis.
          isActive: true
          loginTime: Atque sed occaecati optio.
          mobile: Nihil sed eaque consequuntur.
          nickname: Quasi molestiae.
          username: Eum et repellat itaque itaque delectus.
      errcode: 0
      errmsg: ""
    required:
    - errcode
    - errmsg
  UserLoginByUsernameBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: LoginByUsername_bad_request_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserLoginByUsernameInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: LoginByUsername_internal_server_error_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserLoginByUsernameRequestBody:
    title: UserLoginByUsernameRequestBody
    type: object
    properties:
      captchaId:
        type: string
        description: 图形验证码ID
        default: ""
        example: udf
        minLength: 1
        maxLength: 128
      humanCode:
        type: string
        description: 图形验证码
        default: ""
        example: efz
        minLength: 4
        maxLength: 8
      password:
        type: string
        description: 密码
        example: password
        minLength: 1
        maxLength: 128
      username:
        type: string
        description: 用户名
        example: user
        minLength: 1
        maxLength: 128
    example:
      captchaId: cx4
      humanCode: kt1
      password: password
      username: user
    required:
    - username
    - password
  UserLoginByUsernameResponseBody:
    title: UserLoginByUsernameResponseBody
    type: object
    properties:
      data:
        $ref: '#/definitions/SessionResponseBody'
      errcode:
        type: integer
        description: 错误码
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
    example:
      data:
        cookie: Eos optio ea voluptas voluptatem magnam nostrum.
        credentials:
          expires_in: 25200
          token: eyJhbGciOiJIUz...
        user:
          id: Ipsam magni quia minus officiis.
          isActive: true
          loginTime: Atque sed occaecati optio.
          mobile: Nihil sed eaque consequuntur.
          nickname: Quasi molestiae.
          username: Eum et repellat itaque itaque delectus.
      errcode: 0
      errmsg: ""
    required:
    - errcode
    - errmsg
...
  UserSendSmsCodeBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: SendSmsCode_bad_request_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserSendSmsCodeInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: SendSmsCode_internal_server_error_Response_Body result type (default
      view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserSendSmsCodeRequestBody:
    title: UserSendSmsCodeRequestBody
    type: object
    properties:
      captchaId:
        type: string
        description: 图形验证码ID
        example: i
        minLength: 1
        maxLength: 128
      humanCode:
        type: string
        description: 图形验证码
        example: 4ee
        minLength: 4
        maxLength: 8
      mobile:
        type: string
        description: 手机号
        example: 0a3
        minLength: 11
        maxLength: 11
    example:
      captchaId: wu
      humanCode: p0c
      mobile: skl
    required:
    - mobile
    - humanCode
    - captchaId
  UserSendSmsCodeResponseBody:
    title: UserSendSmsCodeResponseBody
    type: object
    properties:
      data:
        $ref: '#/definitions/SuccessResultResponseBody'
      errcode:
        type: integer
        description: 错误码
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
    example:
      data:
        ok: true
      errcode: 0
      errmsg: ""
    required:
    - errcode
    - errmsg
...
  UserUpdatePasswordRequestBody:
    title: UserUpdatePasswordRequestBody
    type: object
    properties:
      new_password:
        type: string
        example: abc123
        minLength: 6
        maxLength: 128
      old_password:
        type: string
        example: 123abc
        minLength: 1
        maxLength: 128
    example:
      new_password: abc123
      old_password: 123abc
    required:
    - old_password
    - new_password
...
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: |-
      JWT 认证

      **Security Scopes**:
        * `role:user`: 用户
        * `role:admin`: 管理员
    name: Authorization
    in: header